change batt to decimal bsonType

+++ src_server/services/devices/devices.service.js
@@ -33,7 +33,7 @@ module.exports = function (app) {
             bsonType: "int"
           },
           "batt": {
-            bsonType: "int"
+            bsonType: "decimal"
           },
           "type": {
             bsonType: "string"
